requirements.txt
beautifulsoup4==4.12.3
jsonlines==4.0.0
openai==1.58.1
python-dotenv==1.0.1
Requests==2.32.3
toml==0.10.2
tqdm==4.67.1
